# swagger/swagger.yaml
swagger: '2.0'
info:
  description: ''
  version: '2'
  title: MVC
host: 'localhost:2199'
tags:

  - name: Twitter

schemes:
  - http
paths:
  /api/create:
    post:
      tags:
        - Data
      summary: Create DAta
      description: สร้างได้5คำ แล้วเป็นตัวใหญ่
      produces:
        - application/json
      responses:
        '200':
          description: update success
          schema:
            $ref: '#/definitions/createsuccess200'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/badRequest400'
        '401':
          description: unAuthenticate
          schema:
            $ref: '#/definitions/unAuthenticate401'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal500'
  /api/sumtag:
    get:
      tags:
        - Data
      summary: Sumtag
      description: sum random+hastag
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      responses:
        '200':
          description: update success
          schema:
            $ref: '#/definitions/createsuccess200'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/badRequest400'
        '401':
          description: unAuthenticate
          schema:
            $ref: '#/definitions/unAuthenticate401'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal500'
  /api/io:
    get:
      tags:
        - Data
      summary: Get Name have IO
      description: Getname haveIO
      responses:
        '200':
          description: update success
          schema:
            $ref: '#/definitions/createsuccess200'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/badRequest400'
        '401':
          description: unAuthenticate
          schema:
            $ref: '#/definitions/unAuthenticate401'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal500'
definitions:
  createsuccess200:
    type: object
    properties:
      statusCode:
        type: number
        example: 200
      message:
        type: string
        example: update success
  badRequest400:
    type: object
    properties:
      statusCode:
        type: number
        example: 400
      message:
        type: string
        example: Bad Request
  unAuthenticate401:
    type: object
    properties:
      statusCode:
        type: number
        example: 401
      message:
        type: string
        example: Unauthorized
  internal500:
    type: object
    properties:
      statusCode:
        type: number
        example: 500
      message:
        type: string
        example: Internal Server Error